feat(native): set window alpha

+++ crates/ipc/src/subscribe.rs
@@ -19,7 +19,7 @@ pub struct Data {
 
 #[derive(Serialize, Deserialize, Debug)]
 pub struct Container {
-    pub handle: i32,
+    pub handle: isize,
     id: String,
 }
